make interactions model better
