+++ src/types/channel.ts
@@ -5,8 +5,9 @@ import type { Role } from '../structures/role.ts'
 import type { Permissions } from '../utils/permissions.ts'
 import type { EmojiPayload } from './emoji.ts'
 import type { MemberPayload } from './guild.ts'
-import type { InteractionType } from './slash.ts'
+import type { InteractionType } from './interactions.ts'
 import type { UserPayload } from './user.ts'
+import type { MessageComponentPayload } from './messageComponents.ts'
 
 export interface ChannelPayload {
   id: string
@@ -210,6 +211,7 @@ export interface MessageOptions {
   files?: MessageAttachment[]
   allowedMentions?: AllowedMentionsPayload
   reply?: Message | MessageReference | string
+  components?: MessageComponentPayload[]
 }
 
 export interface ChannelMention {
@@ -401,6 +403,7 @@ export interface CreateMessagePayload extends EditMessagePayload {
   message_reference?: MessageReference
   file?: MessageAttachment
   files?: MessageAttachment[]
+  components?: MessageComponentPayload[]
 }
 
 export interface CreateWebhookMessageBasePayload {
